Move and improve slightly the ResourceInputStream wrapper from EpubServer + some more clean up
